requirements.txt
django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
requests==2.22.0
pandas==1.3.4
drf-yasg==1.20.0